update ChefSpec to 4.2, updates spec tests
In 4.1, a slightly updated syntax requires a choice of SoloRunner vs
ServerRunner. Using SoloRunner is the faster test utility.

Adds a few shared examples, and refactors the use cases to be more generalized
and behavior more shared, so the combination of platforms, versions, and
attributes can be tested.

This could be improved further.