Render jmx.yaml template based on data input
After multiple passes at attempting to cover all cases of inclusion of instance
details into a given YAML file, this lays the foundation for leaving the details
up to the end-user to implement the instance hash, instead of trying to capture
each possible flag.
This should help future-proof the recipe and template, so any new flags can be
appended as data.

See #88, #93 and fixes #116, #121.

Tests and examples updated.