convert haproxy integration test from bats to serverspec
When testing with lightweight, limited test environments, the python
yaml package is not available.

While we could retool the test to use the Agent's PYTHONPATH and yaml
package, this pattern is cleaner and provides more valuable debugging
information in the event of a failure.

Other tests will be ported to serverspec/json_spec following.