Add cookbook attribute to monitor resource, and pass to template in create action.
This allows the datadog_monitor resource to be used in other cookbooks. Currently it fails as the template resource in the create action does not search for templates in the datadog cookbook.
Defining this as an attribute with  as the default also allows custom templates to be passed from user cookbooks.