[why-run] Skip dd-handler recipe in why-run mode
Running in why-run mode fails when the chef-handler-datadog gem is not
installed because of the `require 'chef/handler/datadog'` line, so we
skip the entire recipe in why-run mode.

Upgrading chef_handler to v1.2 (Chef>=12)  will allow us to remove that
`require` and unskip the recipe in why-run mode.

See #118