Conditionally use HTTPS based on platform_version
Following #225.

CentOS 5 and earlier use an older version of yum, which embeds an older
version of M2Crypto, that when used against a modern cipher set (i.e. no
sslv3), fails when trying to retrieve the signing key with:

    ...
      File "/usr/lib64/python2.4/site-packages/M2Crypto/SSL/Connection.py",
line 167, in connect_ssl
        return m2.ssl_connect(self.ssl, self._timeout)
    M2Crypto.SSL.SSLError: sslv3 alert handshake failure

There doesn't seem to be any patches to update this behavior, so instead
switch on SSL for any system older than 6.x, continue to use http to
retrieve the key.