Update Kafka configuration to 0.8.2 format
This changes the Chef template to render a file that closely aligns with
the example configuration provided with the dd-agent code. [1] More
details on the changes in the configuration file can be found the
dd-agent repository.

With this change, the rendered file will only correctly collect metrics
on brokers that are on 0.8.2 or higher.

[1] https://github.com/DataDog/dd-agent/pull/2079