Adding spec for datadog::kafka recipe
The kafka recipe didn't yet have a spec associated with it, so this
commit adds the initial spec.

I followed the solr_spec as an example. One change from that example is
I didn't include the fully rendered YAML file, instead preferring to
load the actual file and make assertions against portions of the file
contents.