[yum][repository] Import new RPM key
Ease transition to new RPM key that will sign our RPM packages in the
future.

This will install the new key. The location of the key is defined in
an attribute for users that host the keys internally. The signature of
the downloaded key is checked to make sure that the key is legitimate.

Once packages start being signed with the new key, we should make the
cookbook set the correct `gpgkey` on the `yum_repository` resource
depending on the version of the Agent that needs installing.
Optionally, we could make the cookbook delete the other key.