Simplify postgres.yaml template
Converts the postgres template to using the "simpler" rendering
solution, instead of manually unpacking positional values in YAML.

Using in-template replacement logic, we can modify the values in-place
at the time of rendering.
This does increase memory overhead slightly, as the `instances` object
is copied once, as it starts out as a `Chef::ImmutableArray` that prevents
modification.

- Adds spec tests for specific uses brought up in #379.
- Updates in-recipe documentation to show that one can use either
`server` or `host` values
- Removes a breaking change TODO

Resolves #379

Signed-off-by: Mike Fiedler <miketheman@gmail.com>