Prepare pkg keys rotation (#365)
* [apt][repository] Trust new apt key

Ease transition to new apt key that will sign our repo in the future.

This ensures that the cookbook installs the new key. When we switch to
using the new key to sign the `Release` file of the apt repo, this will
make sure that the new key is trusted.

After the switch, we should make this new key the main key in the
`apt_repository` resource and remove this `execute` resource.
Optionally, we could make the cookbook delete the old key.

* [yum][repository] Import new RPM key

Ease transition to new RPM key that will sign our RPM packages in the
future.

This will install the new key. The location of the key is defined in
an attribute for users that host the keys internally. The signature of
the downloaded key is checked to make sure that the key is legitimate.

Once packages start being signed with the new key, we should make the
cookbook set the correct `gpgkey` on the `yum_repository` resource
depending on the version of the Agent that needs installing.
Optionally, we could make the cookbook delete the other key.

* [ci] Set Ohai log level to :warn

Otherwise `rake spec` can be spammed by debug-level log messages