[trace-agent] Let the trace-agent use its own default settings (#433)
* [trace-agent] Leave the default enabled/disabled behavior to the agent

Since agent 5.13.0, the trace agent is enabled by default.

This cookbook should allow explicitly enabling or disabling the trace
agent, however, when the attribute that controls this is not set, the
cookbook should leave it to the agent to decide whether to run the
trace agent.

This change modifies the default value of an attribute. Technically
this is a breaking change that requires a major version bump of the
cookbook, however I think this change actually improves backward
compatibility and can go with a minor version of the cookbook.

* [trace-agent] Don't set any value in conf unless attribute is overriden

Unless specified by the user, all the conf values of the trace-agent
should be left to their defaults and the trace-agent should choose its
own defaults, not the present cookbook.

Also, overriden trace-agent config attributes will be applied even
if the trace agent is not explicitly enabled.