[service] Avoid failures of `service` resource with frequent restarts (#469)
Happens on systemd-based systems since systemd applies its limits
on the number of restarts of a service (by default, 5 times starts
every 10 seconds) on both the user-requested restarts and the ones
systemd does on its own (when the service fails starting for instance).

Root of the problem is that the service resource in `datadog_monitor`
is different from the one in the main chef run (chef limitation), so
the restarts that happen there are done immediately instead of being
queued up nicely at the end of the global chef run. If multiple
invocations of this resource are updated in a chef run the restart
limit of systemd can be quickly reached.

A better fix would be to remove the service definition from
`datadog_monitor` and make all invocations of `datadog_monitor`
notify the global `service[datadog-agent]` resource. This would be a
breaking change, let's do it for the next major version.