Add rake task to generate monitors for integrations (#720)
It adds a parser module for integrations specifications so that we can now parse them in Ruby. It uses this module to create monitors for the integration. It adds the relevant rake task to write a single or all integrations.

To run on a specific integration:

  bundle exec rake 'create_integration_monitor[<INTEGRATION NAME IN SNAKE CASE>]'

To run on all integrations:

  bundle exec rake create_integration_monitor