Feature: allow muting a host during an MSI install (#778)
MSI installs can take a long time, during which the Agent is not running.
On upgrades, this could trigger no-data monitors.

This PR adds a new option to mute a host while an MSI install takes place,
thus preventing this issue. The host will unmute after either the install
completes, the chef run fails, or after the mute expires (in 60 minutes).

It uses the Datadog API, which requires an application_key.