Removed chef_handler cookbook as its EOL
See https://github.com/chef-boneyard/chef_handler#deprecation:

The chef_handler resource from this cookbook is now shipping as part of Chef 14. With the inclusion of this resource into Chef itself we are now deprecating this cookbook. It will continue to function for Chef 13 users, but will not be updated.