Add unit types to Systemd integration (#859)
Initially the pull request was created to add the functionality to timer units only. I realized it's best to add the possibility to add the functionality for all systemd unit types.

From my point of view it makes the most sense to let the user select the unit type instead of handing over every string added to node attributes in systemd.rb. Therefore an array with all unit types checks if the unit type has been added to node attributes inside the recipe as part of a loop. It could be a little bit confusing for the end user but in the end it is way better to have an efficient code in place.

Co-authored-by: Julien Lebot <julien.ar.lebot@gmail.com>