Add configuration for gearmand integration (#860)
This PR enables the opportunity to handle the gearmand integration via Chef. It's possible to define every parameter described within the integration documentation example file.

Because all parameters for the gearmand integration are optional, the possibility of optional node attributes has been enabled within the configuration deployment, too. Means, if a parameter is not defined, it will be written with default parameters. Well known that the example config file doesn't need that, because these parameters are set by default behind the scenes, the behavior has been set anyway in case of visibility. So the user has the chance to see what he does.

If the user would only add the following to his recipe node.default['datadog']['gearmand']['instances'] = [{}], the basic standard config would be written as well. If only a view parameters will not be set but others, too - the default settings will be written in case of visibility as well.